[#73] Moved logos into icons folder
