+++ src/assets/icons/logos/CHI.js
