+++ src/assets/icons/logos/CLE.js
