+++ src/assets/icons/logos/DAL.js
