+++ src/assets/icons/logos/DEN.js
