+++ src/assets/icons/logos/DET.js
