+++ src/assets/icons/logos/GSW.js
