+++ src/assets/icons/logos/LAL.js
