+++ src/assets/icons/logos/MEM.js
