+++ src/assets/icons/logos/MIA.js
