+++ src/assets/icons/logos/MIN.js
