+++ src/assets/icons/logos/NOP.js
