+++ src/assets/icons/logos/NYK.js
