+++ src/assets/icons/logos/OKC.js
