+++ src/assets/icons/logos/WAS.js
