+++ src/assets/icons/logos/index.js
